motd.template: motd.template: updates and corrections
https://wiki.ghostbsd.org/index.php/FAQ is outdated, and no longer the
preferred page. Remove this line.

Use the new URL for documentation. 'Frequently Asked Questions' is
prominent in the sidebar.

Use the true URL for the list of FreeBSD security advisories.

https://www.GhostBSD.org/releases/ is not the place for release notes
and errata; the page does not exist. Instead, use the development
management page in GitHub.

Add the GhostBSD home page.

To the main branch, from releng/14.2. Closes:

https://github.com/ghostbsd/ghostbsd-src/pull/357